Сделал собственный DI
