+++ app/build.gradle
@@ -58,4 +58,6 @@ dependencies {
     implementation 'io.reactivex.rxjava3:rxandroid:3.0.0'
     implementation "io.reactivex.rxjava3:rxjava:3.1.5"
     implementation 'io.reactivex.rxjava3:rxkotlin:3.0.1'
+
+    implementation project(path: ':diLib')
 }
