+++ app/src/main/java/ru/gbpractice/githapp/di/DiModule.kt
@@ -1,0 +1,33 @@
+package ru.gbpractice.githapp.di
+
+import retrofit2.Retrofit
+import retrofit2.adapter.rxjava3.RxJava3CallAdapterFactory
+import retrofit2.converter.gson.GsonConverterFactory
+import ru.gbpractice.dilib.DiTool
+import ru.gbpractice.githapp.data.RetrofitUserDetailsRepoImpl
+import ru.gbpractice.githapp.data.RetrofitUserListRepoImpl
+import ru.gbpractice.githapp.data.retrofit.GitHubAPI
+import ru.gbpractice.githapp.domain.repos.UserDetailsRepo
+import ru.gbpractice.githapp.domain.repos.UserListRepo
+
+class DiModule() {
+
+    private val baseUrl = "https://api.github.com/"
+
+    private val gitAPI: GitHubAPI by lazy {
+        val adapter = Retrofit.Builder()
+            .baseUrl(baseUrl)
+            .addConverterFactory(GsonConverterFactory.create())
+            .addCallAdapterFactory(RxJava3CallAdapterFactory.create())
+            .build()
+        adapter.create(GitHubAPI::class.java)
+    }
+
+    private val userListRepo: UserListRepo by lazy { RetrofitUserListRepoImpl(gitAPI) }
+    private val userDetailsRepo: UserDetailsRepo by lazy { RetrofitUserDetailsRepoImpl(gitAPI) }
+
+    init {
+        DiTool.add(userListRepo)
+        DiTool.add(userDetailsRepo)
+    }
+}
