+++ settings.gradle
@@ -8,3 +8,4 @@ dependencyResolutionManagement {
 }
 rootProject.name = "GitHApp"
 include ':app'
+include ':diLib'
